3.2 i3334

+++ Admin/Services/DynacaseConsistency/Class.CheckDb.php
@@ -19,7 +19,7 @@
 class checkDb
 {
     /**
-     * @var ressource
+     * @var resource $r
      */
     private $r;
     private $connect;
@@ -160,7 +160,7 @@ class checkDb
         $nb+= pg_num_rows($result);
         $result = pg_query($this->r, "SELECT * from permission where id_application not in (select id from application);");
         $nb+= pg_num_rows($result);
-        
+        $msg='';
         if ($nb > 0) $msg = sprintf("%d unreferenced permissions", ($nb));
         $this->tout["unreferenced permission"] = array(
             "status" => ($nb == 0) ? self::OK : self::BOF,
@@ -386,7 +386,7 @@ class checkDb
             }
             foreach ($oas as $oa) {
                 $aid = $oa->id;
-                if (($oa->docid == $fam->id) && ($oa->type != "array") && ($oa->type != "frame") && ($oa->type != "tab") && ($oa->type != "menu")) {
+                if (($oa->docid == $fam->id) && ($oa->type != "array") && ($oa->type != "frame") && ($oa->type != "tab") && ($oa->type != "") && ($oa->type != "menu")) {
                     $err = self::verifyDbAttr($oa, $pgtype[$aid], $rtype);
                     if ($err) {
                         $cr[] = sprintf("family %s, %s (%s) : %s\n", $fam->getTitle() , $aid, $oa->type, $err) . sprintf("\ttry : drop view family.%s; \n", strtolower($fam->name)) . sprintf("\tand : alter table doc%d alter column %s type %s using %s::%s; \n", $fam->id, $aid, $rtype, $aid, $rtype);
